consolidate controller and views logic
