--- app/controllers/groups/samples/deletions_controller.rb
@@ -1,66 +1,0 @@
-# frozen_string_literal: true
-
-module Groups
-  module Samples
-    # Controller actions for Project Samples Deletions
-    class DeletionsController < Groups::ApplicationController
-      before_action :group, :ensure_enabled
-      before_action :new_dialog_paths, only: :new
-
-      def new
-        authorize! @group, to: :destroy_sample?
-        render turbo_stream: turbo_stream.update('samples_dialog',
-                                                 partial: 'shared/samples/destroy_multiple_confirmation_dialog',
-                                                 locals: {
-                                                   open: true
-                                                 }), status: :ok
-      end
-
-      def destroy_multiple
-        samples_to_delete_count = destroy_samples_params['sample_ids'].count
-
-        deleted_samples_count = Groups::Samples::DestroyService.new(@group, current_user,
-                                                                    destroy_samples_params).execute
-
-        # No selected samples deleted
-        if deleted_samples_count.zero?
-          flash[:error] = t('shared.samples.destroy_multiple.no_deleted_samples')
-        # Partial sample deletion
-        elsif deleted_samples_count.positive? && deleted_samples_count != samples_to_delete_count
-          set_multi_status_destroy_multiple_message(deleted_samples_count, samples_to_delete_count)
-        # All samples deleted successfully
-        else
-          flash[:success] = t('shared.samples.destroy_multiple.success')
-        end
-
-        redirect_to group_samples_path, status: :see_other
-      end
-
-      private
-
-      def group
-        @group = Group.find_by_full_path(params[:group_id]) # rubocop:disable Rails/DynamicFindBy
-      end
-
-      def ensure_enabled
-        not_found unless Flipper.enabled?(:group_samples_destroy)
-      end
-
-      def new_dialog_paths
-        @list_path = list_group_samples_path(list_class: 'sample')
-        @destroy_path = destroy_multiple_group_samples_deletion_path(@group)
-      end
-
-      def destroy_samples_params
-        params.expect(destroy_samples: [sample_ids: []])
-      end
-
-      def set_multi_status_destroy_multiple_message(deleted_samples_count, samples_to_delete_count)
-        flash[:success] = t('shared.samples.destroy_multiple.partial_success',
-                            deleted: "#{deleted_samples_count}/#{samples_to_delete_count}")
-        flash[:error] = t('shared.samples.destroy_multiple.partial_error',
-                          not_deleted: "#{samples_to_delete_count - deleted_samples_count}/#{samples_to_delete_count}")
-      end
-    end
-  end
-end
