+++ app/controllers/samples/deletions_controller.rb
@@ -6,15 +6,29 @@ module Samples
     include ListActions
     respond_to :turbo_stream
 
-    before_action :namespace, only: %i[new destroy_multiple]
-    before_action :confirmation_dialog, only: :new
+    before_action :namespace, only: %i[new destroy destroy_multiple]
+    before_action :confirmation_parameters, only: :new
+    before_action :sample, only: %i[new destroy]
 
     def new
       authorize! (@namespace.group_namespace? ? @namespace : @namespace.project), to: :destroy_sample?
     end
 
+    def destroy
+      project_destroy_service({ sample: @sample })
+
+      if @sample.deleted?
+        flash[:success] = t('.success', sample_name: @sample.name, project_name: @namespace.human_name)
+        redirect_to project_redirect_path
+      else
+        render status: :unprocessable_entity,
+               locals: { type: 'alert', message: error_message(@sample) }
+
+      end
+    end
+
     def destroy_multiple
-      samples_to_delete_count = destroy_multiple_params['sample_ids'].count
+      samples_to_delete_count = destroy_samples_params['sample_ids'].count
 
       deleted_samples_count = destroy_service
 
@@ -38,12 +52,16 @@ module Samples
       @namespace = Namespace.find_by(id: params[:namespace_id])
     end
 
-    def confirmation_dialog
+    def confirmation_parameters
       @confirmation_dialog = "destroy_#{params[:deletion_type]}_confirmation_dialog"
     end
 
-    def destroy_multiple_params
-      params.expect(destroy_multiple: { sample_ids: [] })
+    def sample
+      @sample = Sample.find_by(id: params[:sample_id])
+    end
+
+    def destroy_samples_params
+      params.expect(destroy_samples: [:sample_id, { sample_ids: [] }])
     end
 
     def set_multi_status_destroy_multiple_message(deleted_samples_count, samples_to_delete_count)
@@ -55,18 +73,26 @@ module Samples
 
     def destroy_service
       if @namespace.group_namespace?
-        Groups::Samples::DestroyService.new(@namespace, current_user, destroy_multiple_params).execute
+        Groups::Samples::DestroyService.new(@namespace, current_user, destroy_samples_params).execute
       else
-        Projects::Samples::DestroyService.new(@namespace.project, current_user, destroy_multiple_params).execute
+        project_destroy_service(destroy_samples_params)
       end
     end
 
+    def project_destroy_service(params)
+      Projects::Samples::DestroyService.new(@namespace.project, current_user, params).execute
+    end
+
     def redirect_path
-      if @namespace.group_namespace?
-        group_samples_path(@namespace)
-      else
-        namespace_project_samples_path(@namespace.parent, @namespace.project)
-      end
+      @namespace.group_namespace? ? group_redirect_path : project_redirect_path
+    end
+
+    def group_redirect_path
+      group_samples_path(@namespace)
+    end
+
+    def project_redirect_path
+      namespace_project_samples_path(@namespace.parent, @namespace.project)
     end
   end
 end
