--- app/views/projects/samples/deletions/_new_deletion_dialog.html.erb
@@ -1,31 +1,0 @@
-<%= viral_dialog(open: open, size: :large) do |dialog| %>
-  <% dialog.with_header(title: t(".title")) %>
-  <%= turbo_frame_tag("deletion-alert") %>
-
-  <div
-    data-controller="selection"
-    data-selection-storage-key-value="<%= "#{request.base_url}#{namespace_project_samples_path(@project.namespace.parent, @project)}" %>"
-    class="
-      mb-4 overflow-x-visible font-normal text-slate-500 dark:text-slate-400
-    "
-  >
-    <p class="mb-4">
-      <%= t(".description", sample_name: @sample.name) %>
-    </p>
-    <%= form_for(:deletion, url: namespace_project_samples_deletion_path(sample_id: @sample.id), method: :delete,
-            data: {
-              turbo_frame: "_top",
-              controller: "spinner",
-              action:"turbo:submit-start->spinner#submitStart turbo:submit-end->viral--dialog#close"
-            }
-          ) do |form| %>
-      <%= form.submit t(".submit_button"),
-                  class:
-                    "button text-sm px-5 py-2.5 text-white bg-red-700 border-red-800 hover:bg-red-800 dark:bg-red-600 dark:text-white dark:border-red-600 dark:hover:bg-red-700",
-                  data: {
-                    action: "click->selection#remove",
-                    "selection-id-param": @sample.id,
-                  } %>
-    <% end %>
-  </div>
-<% end %>
