--- app/views/projects/samples/deletions/destroy.turbo_stream.erb
@@ -1,3 +1,0 @@
-<%= turbo_stream.append "flashes" do %>
-  <%= viral_flash(type: type, data: message) %>
-<% end %>
