+++ app/views/samples/deletions/_destroy_multiple_confirmation_dialog.html.erb
@@ -40,7 +40,7 @@
             data: {
               controller: "spinner form--json-submission",
               "form--json-submission-selection-outlet": '#samples-table',
-              "form--json-submission-field-name-value": "destroy_multiple[sample_ids][]",
+              "form--json-submission-field-name-value": "destroy_samples[sample_ids][]",
               "form--json-submission-clear-selection-value": "true",
               "form--json-submission-target": "form",
               action:"turbo:submit-start->spinner#submitStart turbo:submit-end->viral--dialog#close",
